Restructure test source folders
