+++ test/src/TestUtilities/GraphicsUtil.cpp
