+++ test/src/TestUtilities/WindowUtil.hpp
@@ -6,7 +6,7 @@
 #ifndef SFML_TESTUTILITIES_WINDOW_HPP
 #define SFML_TESTUTILITIES_WINDOW_HPP
 
-#include "System.hpp"
+#include "SystemUtil.hpp"
 
 #include <SFML/Graphics/Rect.hpp>
