--- test/src/Window.cpp
@@ -1,10 +1,0 @@
-#define CATCH_CONFIG_MAIN
-
-#include <SFML/Graphics/Rect.hpp>
-#include <SFML/System/Vector2.hpp>
-#include "TestUtilities/Window.hpp"
-
-TEST_CASE("Window unit test exists", "[window]")
-{
-    CHECK(TRUE);
-}
